Enable multiprocess for generator
codebrowser does not support multithreading for the current version,
we may take lots of time on single-threaded generator for large project.

The reason why codebrowser born with single-threaded implementation is
that `clang::tooling::ToolInvocation::run()` is not thread safe. That's
the limit of llvm.

However, we can use multiprocess instead of multithreading to address
this issue.

The output on stderr and stdout may be not correct due to concurrent
output from multiple generator processes, it will be fixed in short.